requirements.txt
Django==1.8.4
PyJWT==1.4.0
django-braces==1.8.1
django-cors-headers==1.1.0
django-oauth-toolkit==0.9.0
django-rest-auth==0.5.0
djangorestframework==3.2.3
mongoengine==0.9.0
oauth2-provider==0.0
oauthlib==1.0.1
pycrypto==2.6.1
pymongo==2.8.1
requests==2.8.1
requests-cache==0.4.10
six==1.10.0
tvdb-api==1.10
